Extract createDirServer — agnostic fetch-native file responder (#1225)
**Stacked on #1219.** Turns the preview file route's serving half into a
reusable, agnostic primitive. `createDirServer(root)` is a fetch-native
`(root, relPath, request) → Response` — streaming byte-range
(`206`/`416`), content-type, lexical traversal guard — that knows
nothing about terminals, git, or kolu.

This is the electricity the [prior-art survey in
#1219](https://github.com/juspay/kolu/issues/1222) pointed at: **no
static-serve library fits** (every one pipes to a Node `res`, severing
the Fetch-`Response` + artifact-sdk decorator composition; none takes a
per-request absolute root), so the responder is *extracted in-app*, not
vendored — publishable as `@kolu/serve-dir` when a second consumer
appears.

### The seam

```
index.ts route  ──(inject root)──▶  createDirServer(root).fetch(tail, req) ──▶ Response
                                          │ serveDir.ts (agnostic)
artifact-sdk decorator (downstream) ◀─────┘ rewrites text/html; composes because
                                            the primitive returns a real Response
                                            and omits Content-Length on full 200s
```

| Before | After |
| --- | --- |
| `iframePreviewRoute.ts` owned serving + URL contract + kolu-git guards
| `serveDir.ts` owns serving (agnostic); `iframePreviewRoute.ts` keeps
only the `?v=<mtime>` URL contract |
| `index.ts`: `serveResolvedFile(resolvePreviewPath(root, tail), root,
range)` | `index.ts`: `createDirServer(root).fetch(tail, c.req.raw)` |

### Deliberate decisions (for reviewers)

- **Symlink/realpath guard dropped** (`kolu-git` `assertRealpathUnder`)
— per explicit direction, to keep the primitive agnostic and given a
local-first threat model. **Lexical** traversal rejection
(decode-then-split + `path.relative` containment) remains. It returns
later as an **injected** guard. The dropped behavior is codified by a
`serveDir.test.ts` case (a symlink is now *followed*) so reinstatement
has a red test to flip — not a silent removal.
- **Root stays the git `repoRoot`** — behavior-preserving (the browse
tree, git-status decoration, and diff are all repo-relative). Switching
the injected root to the terminal's `$PWD` (`meta.cwd`) is now a
one-line change in `index.ts`, deliberately deferred as a separate
browse-model decision.
- **`Content-Length` omitted on full 200s** — kept from #1219 (the
decorator rewrites the body downstream); the 206 path still sets it.

### Coverage
`serveDir.test.ts` (49 cases — range parsing, lexical guard,
200/206/416, atomic-replace streaming, the `createDirServer` Response
wrapper) + the existing code-tab preview e2e (iframe-nav, video, html,
image, live-reload) — all green locally.

> No user-visible behavior change — pure server-internal refactor (plus
the documented guard removal).
