+++ README.md
@@ -205,6 +205,7 @@ pnpm monorepo:
 | `packages/solid-pierre/`             | Solid-native wrappers around [`@pierre/trees`](https://www.npmjs.com/package/@pierre/trees) and [`@pierre/diffs`](https://www.npmjs.com/package/@pierre/diffs); encapsulates Pierre's imperative mount/render lifecycle behind `<FileTree>` and `<CodeView>` with required `onError` props. `<CodeView>` (Pierre's 1.2.x advanced-mode viewport) hosts files and/or diffs in one virtualized scroll — windowed-rendering is unconditional, so single-file callers ride the same path as future multi-file ones |
 | `packages/solid-markdown/`           | App-agnostic safe Markdown → SolidJS renderer: [`marked`](https://marked.js.org/) (GFM + `marked-footnote` / `marked-alert` / `marked-gfm-heading-id`, plus a front-matter strip) parses to HTML (`render.ts`, kept DOM-free + unit-tested, purely structural — it emits marked's default anchors/images and applies no link/image policy), then [DOMPurify](https://github.com/cure53/DOMPurify) sanitizes it (`sanitize.ts`) against a tight Markdown-only allowlist — namespacing every id so a document's anchors can't collide with the app's, and applying an injected `resolveImageSrc` so a repo-relative image loads from the host's file route — (no inline `style`/`class`, SVG/MathML, or form controls) and owns the whole link policy (`safeHref` allowlist; a repo-relative href is tagged for in-app interception so the host opens it in the Code tab, a genuine external href gets new-tab/severed-opener) over *every* anchor, markdown- or inline-HTML-sourced alike — plus an Obsidian-style `[[wikilink]]` inline extension (`markedWikilink`) that mints a distinct `data-md-wikilink` anchor the host resolves pathless across the repo (the document variant only) — so real inline HTML renders (not escaped) while scripts/styles/frames are stripped and links can't hijack the tab. The raw-HTML + image surface is scoped per variant: only `document` (the full-pane preview) gets it; the `inline` / `compact` intent variants keep the stricter scope (no raw block HTML or images), since those are clickable UI rows rendering user/agent text. The output drops into a themed `.kolu-md` container (`markdown.css`, `currentColor` + `color-mix` so it follows light/dark with no theme prop). One `<Markdown>` component with `inline` / `compact` / `document` variants bundling parse mode + styling scale. Kolu's intent surface consumes it, and the `document` variant backs the Code-tab's Markdown rendered-appliance (`@kolu/solid-fileview/renderers/markdown`) that powers the Source ⇄ Rendered preview |
 | `packages/solid-fileview/`           | App-agnostic file-preview "outlet". `<FileView>` owns the Source ⇄ Rendered toggle, mode-availability logic (offered iff a file has *both* a source and a rendered form), and the renderer-registry pick — the core has **no rendering dependencies of its own**: the source view and every rendered form are injected `Renderer` values, and each concrete renderer is a separate `/renderers` sub-path appliance. Ships generic `image` / `video` / `iframe` / `markdown` appliances (the `markdown` one wires `@kolu/solid-markdown`). Kolu's Code-tab plugs in pierre-source + markdown + image + video + an iframe wrapped with the artifact-sdk comment bridge, so `right-panel/BrowseFileDispatcher` is a thin `fsReadFile`→`FileView` adapter rather than bespoke render wiring; `.md` now shows the toggle (rendered by default) |
+| `packages/serve-dir/`                | App-agnostic `@kolu/serve-dir` — a fetch-native directory file server: `createDirServer(root, realpathGuard?)` returns `fetch(relPath, request) → Response` with streaming HTTP byte-range (`206`/`416`), per-extension Content-Type, and a lexical traversal guard. **Zero workspace deps** (only `node:fs`/`path`/`stream`); knows nothing about terminals/git/kolu, so the dependency arrow points *out* (`kolu-server → @kolu/serve-dir`). The consumer injects the absolute root and an optional realpath/symlink guard (kolu-server wires kolu-git's `assertRealpathUnder`), and composes any response transform (the artifact-sdk `<script>` decorator) downstream — which works precisely because `fetch` returns a real `Response` and omits `Content-Length` on full 200s. Backs the Code-tab iframe/image/video preview file route |
 | `packages/solid-browser/`            | App-agnostic, framework-free navigation layer for the Code tab — the *browser* that moves *between* documents while `@kolu/solid-fileview` draws each one. Phase 1 (shipped) is pure path/location math with a single workspace dep (`solid-browser → @kolu/solid-markdown/url-policy`); phase 2 adds the `<Browser>` that composes `<FileView>`, extending the acyclic graph to `solid-browser → solid-fileview → solid-markdown`. Ships the pure path math: `resolveRelativePath`/`resolveLinkHref` (GitHub-style relative-ref resolution — resolve against the doc's directory, root-absolute from the repo root, reject own-scheme/`..`-escape) and `pathFromPreviewPathname` (invert a sandboxed iframe's reported `location.pathname` back to a document path, via a host-injected `PreviewPathCodec` so the inversion can't drift from the encoding). Its only dep is `@kolu/solid-markdown/url-policy` (the DOM-free `hasOwnScheme` shape test, shared with the markdown href policy) — zero kolu imports, so the dependency arrow points *out* of the client. The client consumes it from `markdownImageSrc` (which wraps `resolveRelativePath` into a per-terminal file-route URL), `BrowseFileDispatcher` (which calls `resolveLinkHref` before opening a relative link in the Code tab), and `BrowseIframeRenderer` (which binds kolu's `encodePreviewPath`/`decodePreviewPath` codec into `pathFromPreviewPathname`). Phase 2 adds `createBrowser` (location + history) + a `<Browser>` composing `<FileView>` so back/forward falls out for free; see the [solid-browser Atlas note](docs/atlas/dist/solid-browser.html) |
 | `packages/artifact-sdk/`             | Self-contained comments-on-files toolkit. Three exports: `./core` (pure W3C TextQuoteSelector functions used by both runtimes), `./client` (parent-side iframe ↔ parent bridge + core re-exports), `./server` (one-line `mountArtifactSdk(app, ...)` that wires the SDK bundle route and an HTML-decoration middleware — esbuild bundles the in-iframe script at server startup, hash-keyed for cache busting). The host server never imports HTML rewriting logic |
 | `packages/server/`                   | [Hono](https://hono.dev/); owns `@kolu/pty-host` in-process but consumes it through the typed `ptyHostSurface` contract (an in-process link), and runs the per-terminal provider DAG fresh in kolu-server against its taps                     |
