+++ docs/atlas/src/content/atlas/electricity.mdx
@@ -39,6 +39,7 @@ The live example: asked to add back/forward to the Code-tab preview, a session p
 | **@kolu/log** · **@kolu/html-escape** | Zero-dep leaf types every package can import without dragging the domain tree. | <PrLink pr={1089} /> | <Pill variant="done">done</Pill> |
 | **dom/** + injectable `isMac` | `walkShadowRoots` relocated to a neutral wall; keybind core no longer reads the UA singleton. | <PrLink pr={1089} /> | <Pill variant="done">done</Pill> |
 | **createSharedRoot** | Lazy app-scoped reactive singleton (7 consumers). 100% agnostic SolidJS. | <PrLink pr={974} /> | <Pill variant="done">done</Pill> <Pill variant="run">publish?</Pill> |
+| **@kolu/serve-dir** | Fetch-native file serving from an absolute root — streaming byte-range (`206`/`416`), content-type, lexical traversal guard; a pure `(root, relPath, request) → Response`. Its **own package, zero workspace deps** (only `node:fs`/`path`/`stream`) — the dependency arrow points *out* (`kolu-server → @kolu/serve-dir`). Agnostic of terminals/git/kolu; the consumer injects the root and composes the artifact-sdk `<script>` decorator downstream. A 20-agent prior-art survey found no static-serve library fits (all pipe to a Node `res`, breaking the Fetch-`Response` + decorator composition; none takes a per-request absolute root or does a realpath guard), so it was extracted, not vendored. Path safety is two-stage by volatility: the **lexical** guard (decode-then-split + `path.relative` containment) is built in (pure, universal), while the **realpath/symlink** guard is an *injected* `realpathGuard` the consumer supplies (fs-touching, threat-model-specific) — `kolu-server` wires kolu-git's `assertRealpathUnder`, so the package stays agnostic without dropping the security boundary. A **leaf** (bounded algorithm, not a transport-grade volatility). | <PrLink pr={1225} /> | <Pill variant="done">done</Pill> <Pill variant="run">②/③ proof-pending</Pill> |
 | **@kolu/solid-xterm** (createSolidXterm) | xterm.js lifecycle — 8-addon orchestration, WebGL single-owner + context-loss recovery, FitAddon resize, owner-correct async dispose across `await`. | <PrLink pr={1116} /> (plan) | <Pill variant="fix">to build</Pill> |
 | **@kolu/solid-browser** | The Code tab *is* a browser — location/address model, back-forward history, cross-content-type link interception, GitHub-relative resolution, sandboxed-iframe lifecycle. git is an injected resolver; the renderer packages plug in. [Plan](solid-browser). | <PrLink pr={1191} /> | <Pill variant="run">phase 1</Pill> primitives shipped · createBrowser+history (P2), ③ proof-pending |
 
@@ -97,7 +98,6 @@ Looked framework-sized (the audit's biggest braids), but fail the bar — record
 | **@kolu/surface-publish** | <Pill variant="bad">already-covered</Pill> | Surface's `cellHandlers`/`collectionHandlers`/`pollOnEvent` already own server snapshot+delta. Only `terminal.attach` is a raw escape hatch. |
 | **@kolu/commands** (palette) | <Pill variant="bad">domain-coupled</Pill> | Palette engine is already agnostic; all coupling is isolated in `createCommands`, which must stay in the app. Extracting deletes ~0 kolu lines. |
 | God-procedure split | <Pill variant="run">refactor</Pill> | Its biggest dissolvable chunk *is* `solid-xterm` (above); the rest is kolu orchestration — a refactor, not a receptacle. |
-| **Preview file route** (`serveResolvedFile` / `parseByteRange`, <PrLink pr={1219} />) | <Pill variant="bad">in-app receptacle</Pill> | A 20-agent prior-art survey found **no drop-in fits**, *and* it isn't even a lean-on-external case. Every full static server (`send` / `serve-static` / `@fastify/static` / `koa-send` / `st` / `sirv`) pipes straight to a Node `res` socket — severing **both** the Fetch-`Response` contract *and* the artifact-sdk HTML-decorator composition (the in-iframe-nav regression class: a serving lib pins `Content-Length`, the downstream `<script>` splice can't fix it) — and **none** does a realpath symlink guard, which kolu *needs* for agent-written repos (cf. `@hapi/inert` leaking `/etc/shadow`; CVE-2026-6414 `%2f` bypass vindicates kolu's decode-then-split). `@hono/node-server`'s `serveStatic` additionally can't take an absolute per-terminal root. The only full-fit shape is exactly what the route already does — `createReadStream({start,end})` → `Readable.toWeb` → `Response`, omit `Content-Length` on 200 — independently converged on by Deno `@std/http` and SvelteKit/Vite. Three kolu-only contracts converge here (per-terminal `repoRoot`, decorator composition, symlink threat), so it's a local **receptacle**, not a leaf; `range-parser` would even *regress* RFC-9110 suffix handling. Hand-rolled is justified. |
 
 ---
